Detect whether Linux or FreeBSD shutdown command should be ran

+++ usr/lib/arcologout/arcologout.py
@@ -10,6 +10,7 @@ import Modal
 import Functions as fn
 import threading
 import signal
+import sys
 
 gi.require_version('Gtk', '3.0')
 gi.require_version('Gdk', '3.0')
@@ -19,7 +20,7 @@ from gi.repository import Gtk, GdkPixbuf, Gdk, Wnck, GLib, GdkX11  # noqa
 
 
 class TransparentWindow(Gtk.Window):
-    cmd_shutdown = "sudo shutdown -P now"
+    cmd_shutdown = "sudo shutdown -P now" if sys.platform == "linux" else "sudo shutdown -p now"
     cmd_restart = "sudo shutdown -r now"
     cmd_suspend = "systemctl suspend"
     cmd_hibernate = "systemctl hibernate"
